requirements.txt
fastapi
uvicorn
python-telegram-bot==20.7
python-multipart
python-dotenv
openai
openai-agents
asyncio
apscheduler
yfinance
pydantic